Initial import of the trinity project.
This is a new incarnation (a third one, to be exact) of the educational raytracer;
the successor of retrace and FMIRay.

It is very empty right now, and is almost a direct fork of the initial commits to
"retrace". No project files right now, to be added in a subsequent commit.