Add Code::Blocks project file template, and a .gitignore file.
The .cbp file is a TEMPLATE only; the intended use is to copy this file
to `trinity.cbp' and then open that file with Code::Blocks. This way,
if you need local modifications to the project file, git won't track
it as if they need to be committed (and trinity.cbp is added to gitignore).