Fix the bug with antialiasing detection.
The bug was caused by that the detection of the pixels that need
anti-aliasing, and marking of the "interesting" pixels was done in
a single pass. When a previous pixel was deemed "interesting", the
next one had a neighbour which was pure red, and then it turn it was
detected as if it needs antialiasing as well. This caused the picture
to be cluttered with a red sierpinski-triangle like fractals.

The correct code is employed now. Detection of "interesting" pixels
is done in a pass on its own, with the pixels in question being marked
in the needsAA array. After that, another pass on its own does the
painting of the interesting pixels. What you can see now is the
result of the edge-detection, painted in red.

No adaptive antialiasing is done currently, to be done in the next
commit.