Fix the bug with the Csg operations.
The bug was caused by the detection of whether we're inside or outside
a geometry. The detection solely relies on number of intersection points
being odd/even. However, the findAllIntersections() function only checks
for intersections up to maxDist - so it may miss some points, and the
inL/inR variables will be initialized to the wrong values.

Solution: modify findAllIntersection() to find all intersections, up to
infinity. And move the check for maxDist inside CsgOp::intersect.

Also, add some debugging code to the raytracer. When the rendering
is finished, and you click somewhere on the image, a 'debug' ray is
sent through that pixel, with the information about the intersection
point/geometries printed to stdout.