Add a hack to speed up the rendering of glossy reflections.
Don't evaluate fully a glossy reflection if the ray we're dealing with
isn't a camera ray (depth = 0). If it's somewhere deeper in the pipeline,
shoot out 5 rays only.

This pretty much avoids the combinatorial explosion that the scene
exhibits: when the numSamples of the glossy sphere is increased
to 120, the rendertime increases almost exponentially. Some figures:

120 samples, without the hack: rendertime: 753 seconds (!!)
120 samples, with the hack   : rendertime: 9 seconds.