Correct I/O gamma treatment.
The output image, on transfer from vfb[][] to SDL's buffer, is gamma-compressed
(converted to sRGB).

The input images (on load in BitmapTexture's constructor) are
gamma-decompressed (a sRGB->linear conversion by default; if a particular
parameter is given to the assumedGamma value, and it is not 2.2, the
conversion is a simple power-law curve with the specified gamma value).