Speed up rendering (don't anti-alias overexposed areas).
The old code in the tooDifferent() function looked for color differences
of more than 0.1 in any of the RGB channels - in which case the colors
were deemed "too different" unconditionally, and an AA treatment to
the respective pixel was employed.

This doesn't make sense for the overexposed areas, which happen a lot with
the current scene, because of the HDR background and the reflective/glossy
surfaces. There, the relative difference between neighbouring overexposed
pixels is similar to non-overexposed pixels which don't require AA.
However, since their magnitude is larger, the absolute color difference
is proportionally larger as well, and they almost always qualify for
the AA pass. This results in needless oversampling in the overexposed
areas, which is moot anyway; they are overexposed; there's no detail there.

The new code doesn't consider it interesting when the darker of the two
pixels is more than 1.33 in absolute brightness (technically, 1.0 should
be correct here as well, but let's be a little on the safe side).