Fix ParsedBlock::requiredProp()
This fixes the hiccup in the video to lecture 8
(https://vimeo.com/80786651) at 45:24. The code in requiredProp() calls
a SyntaxError constructor. However, there are two SyntaxError constructors,
one is SyntaxError(int, const char*, const char*), and the other one is
SyntaxError(int, const char*, ...). The intended one to be called is the
second one, but the first is called instead, as it is more specific.

The ambiguity is resolved by removing the first constructor, as it hasn't
actually been used anywhere.