When the mesh has no normals in the .OBJ file, make up some.
If an .OBJ file has no "vn" lines, no per-vertex normals will be
available, and only gnormals of triangles can be used in shading.
This implies that the mesh will always be faceted. We can, however,
generate some normals, by finding the incident triangles to each
vertex, and averaging their gnormals.

Added a parameter, autoSmooth, which controls whether this step
is desirable.