Make RectLight vary its output accoring to cos(shadePoint, lightDir)
Since our RectLight was assumed to be shining in one hemisphere only (-y
canonical direction), there was a sharp boundary between lit and unlit
in boxed.trinity. It makes much more sense to have the light's output
dependant on the angle through which the rect light is "seen" from the
shade point: the more grazing angles we consider, the less the light.
Implement a cos(shadePointDir, lightDir) law. This makes the boundary smooth.