Optimize Color::toRGB32 (which improves rendertime significantly).
This optimization is needed for the interactive mode.
The function displayVFB() uses Color::toRGB32(), which relies on
convertTo8Bit_sRGB, which applied sRGB compression to the output
pixels. The gamma-compressed part of the colorspace required
powf(), which is quite slow if you need to run it against several
million pixels per second. To optimize, cache the results at
4096 points between 0 and 1 and use that lookup table instead of
the true pow().

testmesh.trinity at 320x240, single-threaded, improved from 6.80 to
8.60 fps (~25%) on my machine.