Eliminate some divisions in BBox intersection methods.
Since divisions by ray.dir.(x|y|z) happen a lot, it's beneficial to
compute the reciprocal of ray.dir before the intersection calculation.
This is made with a special, augmented Ray class (RRay).

The same kind of optimization can be implemented in other parts of the
raytracer, but doing it globally makes no sense: RRay::dir and rdir have
to be kept in sync, which is too much work to do (esp. through
model-view transformations).

This optimization sheds about 2.5% of the render time.